requirements.txt
flask
pillow
psycopg2
argon2
bcrypt
cryptography